+++ packages/extension-base/src/background/KoniTypes.ts
@@ -1902,7 +1902,7 @@ export interface Notification {
 export type NotificationParams = Omit<Notification, 'id'>;
 
 export interface CronReloadRequest {
-  data: 'nft' | 'staking'
+  data: 'nft' | 'staking' | 'balance' | 'crowdloan'
 }
 
 export interface AllLogoMap {