+++ packages/extension-base/src/koni/background/subscription.ts
@@ -80,7 +80,8 @@ export class KoniSubscription {
     const currentAddress = this.state.keyringService.currentAccount?.address;
 
     if (currentAddress) {
-      this.subscribeBalancesAndCrowdloans(currentAddress, this.state.getChainInfoMap(), this.state.getChainStateMap(), this.state.getSubstrateApiMap(), this.state.getEvmApiMap());
+      this.subscribeBalances(currentAddress, this.state.getChainInfoMap(), this.state.getChainStateMap(), this.state.getSubstrateApiMap(), this.state.getEvmApiMap());
+      this.subscribeCrowdloans(currentAddress, this.state.getSubstrateApiMap());
       this.subscribeStakingOnChain(currentAddress, this.state.getSubstrateApiMap());
     }
 
@@ -98,7 +99,8 @@ export class KoniSubscription {
         return;
       }
 
-      this.subscribeBalancesAndCrowdloans(address, serviceInfo.chainInfoMap, serviceInfo.chainStateMap, serviceInfo.chainApiMap.substrate, serviceInfo.chainApiMap.evm);
+      this.subscribeBalances(address, serviceInfo.chainInfoMap, serviceInfo.chainStateMap, serviceInfo.chainApiMap.substrate, serviceInfo.chainApiMap.evm);
+      this.subscribeCrowdloans(address, serviceInfo.chainApiMap.substrate);
       this.subscribeStakingOnChain(address, serviceInfo.chainApiMap.substrate);
     };
 
@@ -116,17 +118,28 @@ export class KoniSubscription {
     return Promise.resolve();
   }
 
-  subscribeBalancesAndCrowdloans (address: string, chainInfoMap: Record<string, _ChainInfo>, chainStateMap: Record<string, _ChainState>, substrateApiMap: Record<string, _SubstrateApi>, web3ApiMap: Record<string, _EvmApi>, onlyRunOnFirstTime?: boolean) {
-    this.state.handleSwitchAccount(address).then(() => {
-      const addresses = this.state.getDecodedAddresses(address);
+  subscribeBalances (address: string, chainInfoMap: Record<string, _ChainInfo>, chainStateMap: Record<string, _ChainState>, substrateApiMap: Record<string, _SubstrateApi>, web3ApiMap: Record<string, _EvmApi>, onlyRunOnFirstTime?: boolean) {
+    const addresses = this.state.getDecodedAddresses(address);
 
-      if (!addresses.length) {
-        return;
-      }
+    if (!addresses.length) {
+      return;
+    }
 
+    this.state.handleResetBalance(address).then(() => {
       this.updateSubscription('balance', this.initBalanceSubscription(addresses, chainInfoMap, chainStateMap, substrateApiMap, web3ApiMap, onlyRunOnFirstTime));
-      this.updateSubscription('crowdloan', this.initCrowdloanSubscription(addresses, substrateApiMap, onlyRunOnFirstTime));
     }).catch((err) => this.logger.warn(err));
+  }
+
+  subscribeCrowdloans (address: string, substrateApiMap: Record<string, _SubstrateApi>, onlyRunOnFirstTime?: boolean) {
+    const addresses = this.state.getDecodedAddresses(address);
+
+    if (!addresses.length) {
+      return;
+    }
+
+    this.state.resetCrowdloanMap(address).then(() => {
+      this.updateSubscription('crowdloan', this.initCrowdloanSubscription(addresses, substrateApiMap, onlyRunOnFirstTime));
+    }).catch(console.error);
   }
 
   subscribeStakingOnChain (address: string, substrateApiMap: Record<string, _SubstrateApi>, onlyRunOnFirstTime?: boolean) {
@@ -330,6 +343,23 @@ export class KoniSubscription {
     const currentAddress = this.state.keyringService.currentAccount?.address;
 
     this.subscribeStakingOnChain(currentAddress, this.state.getSubstrateApiMap());
+
+    await waitTimeout(1800);
+  }
+
+  async reloadBalance () {
+    const currentAddress = this.state.keyringService.currentAccount?.address;
+
+    await this.state.handleResetBalance(currentAddress, true);
+    this.subscribeBalances(currentAddress, this.state.getChainInfoMap(), this.state.getChainStateMap(), this.state.getSubstrateApiMap(), this.state.getEvmApiMap());
+
+    await waitTimeout(1800);
+  }
+
+  async reloadCrowdloan () {
+    const currentAddress = this.state.keyringService.currentAccount?.address;
+
+    this.subscribeCrowdloans(currentAddress, this.state.getSubstrateApiMap());
 
     await waitTimeout(1800);
   }