+++ packages/extension-koni-ui/src/utils/index.ts
@@ -226,3 +226,5 @@ export * from './sort';
 export * from './transaction';
 export * from './walletConnect';
 export * from './crowdloan';
+export * from './crowdloan';
+export * from './history';